pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hat-yai"
version = "0.1.0"
description = "AirSaas Company Audit Agent — LangGraph Cloud"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langsmith>=0.2.0",
    "firecrawl-py>=1.0.0",
    "supabase>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
include = ["hat_yai*"]
